backend/requirements.txt
# backend/requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Redis
redis==5.0.4

# Task queue
celery==5.4.0
flower==2.0.1

# AI / Agent
google-generativeai==0.8.5
langchain==0.2.1
langgraph==0.0.55
langfuse==2.24.0

# Retry logic
tenacity==8.3.0

# Notifications
twilio==9.0.4
sendgrid==6.11.0

# HTTP client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1

# Payment gateway
razorpay==1.4.2
